Made sure to use comfy prefix for comfy imports, renamed sampling_motion to sampling
