+++ animatediff/nodes.py
@@ -10,7 +10,7 @@ from .model_utils import get_available_motion_loras, get_available_motion_models
 from .motion_utils import NoiseType
 from .motion_lora import MotionLoraInfo, MotionLoraList
 from .model_injection import InjectionParams, ModelPatcherAndInjector, MotionModelSettings, load_motion_module
-from .sampling_motion import motion_sample_factory
+from .sampling import motion_sample_factory
 
 from .nodes_extras import AnimateDiffUnload, EmptyLatentImageLarge, CheckpointLoaderSimpleWithNoiseSelect
 from .nodes_experimental import AnimateDiffModelSettingsSimple, AnimateDiffModelSettingsAdvanced, AnimateDiffModelSettingsAdvancedAttnStrengths
