+++ animatediff/nodes_extras.py
@@ -4,6 +4,7 @@ import folder_paths
 import nodes as comfy_nodes
 from comfy.model_patcher import ModelPatcher
 from comfy.sd import load_checkpoint_guess_config
+
 from .logger import logger
 from .model_utils import IsChangedHelper, BetaSchedules
